requirements.txt
# Core data science packages
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter for exploration
jupyter>=1.0.0
ipykernel>=6.25.0

# Web scraping for data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Data processing
openpyxl>=3.1.0
xlrd>=2.0.1

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Model persistence
joblib>=1.3.0

# Optional: Advanced ML (for future use)
# xgboost>=1.7.0
# lightgbm>=4.0.0
# catboost>=1.2.0
